Move mock data files to separate folder for organization
