+++ FoodShare/src/MockData/labeled_input.tsx
